+++ bin/main.rs
@@ -47,12 +47,12 @@ fn validate_machine_id() {
 }
 
 #[cfg(feature = "dev")]
-fn route_cmd(opt: Opt) -> i32 {
+async fn route_cmd(opt: Opt) -> i32 {
     match opt {
         Opt::Codebook(cmd) => routes::codebook(cmd),
         Opt::Bench(cmd) => routes::bench(cmd),
         Opt::Regression(cmd) => regression::regression(cmd),
-        Opt::Run(cmd) => routes::run(cmd),
+        Opt::Run(cmd) => routes::run(cmd).await,
         Opt::Summarize(cmd) => routes::summarize_engine(cmd),
         Opt::RegenExamples(cmd) => routes::regen_examples(cmd),
         Opt::GenerateEncyrptionKey => routes::keygen(),
@@ -60,23 +60,24 @@ fn route_cmd(opt: Opt) -> i32 {
 }
 
 #[cfg(not(feature = "dev"))]
-fn route_cmd(opt: Opt) -> i32 {
+async fn route_cmd(opt: Opt) -> i32 {
     match opt {
         Opt::Codebook(cmd) => routes::codebook(cmd),
-        Opt::Run(cmd) => routes::run(cmd),
+        Opt::Run(cmd) => routes::run(cmd).await,
         Opt::Summarize(cmd) => routes::summarize_engine(cmd),
         Opt::GenerateEncyrptionKey => routes::keygen(),
     }
 }
 
-fn main() {
+#[tokio::main]
+async fn main() {
     validate_machine_id();
 
     env_logger::init();
 
     let opt = Opt::from_args();
 
-    let exit_code = route_cmd(opt);
+    let exit_code = route_cmd(opt).await;
 
     std::process::exit(exit_code);
 }