+++ Changelog.md
@@ -15,6 +15,9 @@
 - Updated Count hyper prior to prevent so many underflow/overflow errors and
     provide better fit
 - Braid automatically converts old metadata into new metadata
+- Engine UpdateInformation using tokio RwLock fo async
+- Engine run progress bars use async for faster running due to non-blocking
+    sleep
 
 ## 0.37.0
 - `Engine::save` now borrows an immutable reference to the engine, thus `save`